psf: widen OutOfBounds.size to u64 and guard the no-default-features build
OutOfBounds.size was usize while offset/length are u64; the reader knows
length as u64, so it had to cast lossily. Widen the field to u64 so both
paths report it without a cast (pre-1.0, so the API break is free).

Add a CI job that runs the suite with --no-default-features: every other
job uses --all-features and would mask a regression in the sans-IO core.

+++ src/lib.rs
@@ -37,11 +37,7 @@ pub enum Error {
     #[error("truncated PSF: need {need} bytes, have {have}")]
     Truncated { need: usize, have: usize },
     #[error("stream range {offset}+{length} out of bounds (file is {size} bytes)")]
-    OutOfBounds {
-        offset: u64,
-        length: u64,
-        size: usize,
-    },
+    OutOfBounds { offset: u64, length: u64, size: u64 },
     #[error("CIX parse error: {0}")]
     Cix(String),
 }
@@ -105,7 +101,7 @@ impl<'a> Psf<'a> {
             .ok_or(Error::OutOfBounds {
                 offset: s.offset,
                 length: s.length,
-                size: self.data.len(),
+                size: self.data.len() as u64,
             })?;
         Ok(&self.data[off..end])
     }
